Split widget: info row
