+++ extension/devtools/app/lib/widgets/info_row.dart
@@ -1,0 +1,22 @@
+part of '../main.dart';
+
+class _InfoRow extends StatelessWidget {
+  const _InfoRow({required this.label, required this.value});
+
+  final String label;
+  final String value;
+
+  @override
+  Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8),
+      child: Row(
+        children: [
+          Expanded(child: Text(label, style: textTheme.bodySmall)),
+          Text(value, style: textTheme.bodySmall),
+        ],
+      ),
+    );
+  }
+}
